refactor(batch): keep safety margin for max gRPC message size
Calculating the size precisely is possible, but requires
writing code that understands the internal structure of
the gRPC messages. A much simpler approach is to estimate
the size approximately and keep a tiny (compared to the
total message size) margin for error.

+++ src/main/java/io/weaviate/client6/v1/api/collections/batch/Batch.java
@@ -74,7 +74,7 @@ final class Batch {
    * This is determined by the {@link GrpcChannelOptions#maxMessageSize}.
    */
   @GuardedBy("this")
-  private final long maxSizeBytes;
+  private final int maxSizeBytes;
 
   /** Total serialized size of the items in the {@link #buffer}. */
   @GuardedBy("this")
